style indoor elements
why does it have to be so hard
i had to edit a stupid lua file just to load my custom nodes with "indoor" tags.
why doesn't it just load everything.
why does it have to be so hard

and then I have to edit a style.json file too!!!