Fix some ui state errors
This is scuffed. Where's my UI library when I need it...

When selecting a starting position, you can't move the pin or get a
new geolocation.

I added a geolocation error popup for when it fails for whatever
reason.

Reset start location and end location names when exiting navigation so
that if you navigate again it doesn't show old values.